Develop

+++ app/Http/Controllers/ContactController.php
@@ -5,27 +5,26 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
 use App\Contact;
+use App\Address;
 use Mapper;
+use View;
 
 class ContactController extends Controller
 {
+    public function __construct()
+    {
+        View::share('URL_HTTPS', env('APP_SECURE'));
+    }
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {  
-        $contacts = Contact::orderBy('firstname', 'asc')->get();
-        //php pagination
-        // ->paginate(10);
-        foreach($contacts as $contact)
-        {
-            if($contact->phone != ''){
-                $contact->phone = Contact::phoneFormat($contact->phone);
-            }
-        }
-        return view('contacts.index', compact('contacts'));
+    {
+        return view('contacts.index')->with([
+            'contacts' => Contact::all()
+        ]);
     }
 
     /**
@@ -46,23 +45,30 @@ class ContactController extends Controller
      */
     public function store(Request $request)
     {
+        //Validation
         $request->validate([
-            'firstname' => 'required|min:2',
-            'lastname' => 'required|min:2',
-            'email'=>'required|unique:contacts'
-            ]);
-        Contact::create([
-            'firstname'=>trim(strtolower(request()->input('firstname'))),
-            'lastname'=>trim(strtolower(request()->input('lastname'))),
-            'email'=>trim(strtolower(request()->input('email'))),
-            'phone'=>trim((request()->input('phone'))),
-            'birthday'=>request()->input('birthday'),
-            'address1'=>trim(strtolower(request()->input('address1'))),
-            'address2'=>trim(strtolower(request()->input('address2'))),
-            'city'=>trim(strtolower(request()->input('city'))),
-            'state'=>request()->input('state'),
-            'zip'=>request()->input('zip'),
+            'firstname' => 'required',
+            'lastname' => 'required',
+            'email' => 'required|email|unique:contacts',
+            'phone' => 'digits:10|nullable',
+            'address1' => 'required_with:zip',
+            'zip' => 'required_with:address1|digits:5|nullable',
         ]);
+
+        $post = $request->all();
+
+        try {
+            //Create contact
+            $contact = Contact::create($post);
+
+            //Create address
+            if (!empty($post['address1']) && !empty($post['zip'])) {
+                $contact->address()->create($post);
+            }
+        } catch (\Exception $e) {
+            abort(500, $e->getMessage());
+        }
+
         session()->flash('message', 'Your contact has been added.');
         return redirect('/');
     }
@@ -74,24 +80,25 @@ class ContactController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function show(Contact $contact)
-    {  
-        $address = $contact->address1 . " " . $contact->address2 . " " . $contact->city . " " . $contact->state . " ". $contact->zip;
-        $prepAddr = str_replace(' ','+',$address);
-        $geocode=file_get_contents('https://maps.google.com/maps/api/geocode/json?address='.$prepAddr.'&sensor=false&key=AIzaSyAjc7C1wVVqG6qUwJTS3d55Nu4xxen1yJ4');
-        $output= json_decode($geocode);
-        if(count($output->results)==0){
-            $lat = "34.0522342";
-            $long = "-118.2436849"; 
-        } else{
+    {
+        try {
+            $address = $contact->address;
 
-            $lat = $output->results[0]->geometry->location->lat;
-            $long = $output->results[0]->geometry->location->lng;
-        }
-        Mapper::map($lat, $long); 
-         if($contact->phone != ''){
+            //Get latitude and longtitude to make google map
+            $location = Address::getLocation($address);
+            Mapper::map($location['lat'], $location['long']);
+
+            //Make phone number pretty
+            if (!empty($contact->phone)) {
                 $contact->phone = Contact::phoneFormat($contact->phone);
             }
-        return view('contacts.show', compact('contact'));
+        } catch (\Exception $e) {
+            abort(500, $e->getMessage());
+        }
+
+        return view('contacts.show')->with([
+            'contact' => $contact
+        ]);
     }
 
     /**
@@ -100,10 +107,11 @@ class ContactController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
-    {    
-        $contact = Contact::find($id);
-        return view('contacts.edit', compact('contact'));
+    public function edit(Contact $contact)
+    {
+        return view('contacts.edit')->with([
+            'contact' => $contact
+        ]);
     }
 
     /**
@@ -113,28 +121,37 @@ class ContactController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, Contact $contact)
     {
+        //Validation
         $request->validate([
-            'firstname' => 'required|min:2',
-            'lastname' => 'required|min:2',
-            'email'=>[
-                    'required',
-                    Rule::unique('contacts')->ignore($id),
-                     ],
+            'firstname' => 'required',
+            'lastname' => 'required',
+            'phone' => 'digits:10|nullable',
+            'email' => [
+                'required',
+                'email',
+                Rule::unique('contacts')->ignore($contact->id),
+            ],
+            'address1' => 'required_with:zip',
+            'zip' => 'required_with:address1|digits:5|nullable',
         ]);
-        $contact = Contact::find($id);
-        $contact->firstname = trim(strtolower(request()->input('firstname')));
-        $contact->lastname = trim(strtolower(request()->input('lastname')));
-        $contact->email = trim(strtolower(request()->input('email')));
-        $contact->phone = trim((request()->input('phone')));
-        $contact->birthday = request()->input('birthday');
-        $contact->address1 = trim(strtolower(request()->input('address1')));
-        $contact->address2 = trim(strtolower(request()->input('address2')));
-        $contact->city = trim(strtolower(request()->input('city')));
-        $contact->state = request()->input('state');
-        $contact->zip = request()->input('zip');
-        $contact->save();
+        $post = $request->all();
+
+        try {
+            //Update Contact
+            $contact->update($post);
+
+            $address = $contact->address;
+
+            //Create or Update address
+            if (!empty($post['address1']) && !empty($post['zip'])) {
+                empty($address) ?  $contact->address()->create($post) : $address->update($post);
+            }
+        } catch (\Exception $e) {
+            abort(500, $e->getMessage());
+        }
+
         session()->flash('message', 'Your contact has been updated.');
         return redirect('/');
     }
@@ -145,11 +162,18 @@ class ContactController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function delete(Request $request)
     {
-        $contact = Contact::find($id);
+        $result = ['status' => 1];
+
+        $contact = Contact::find($request->id);
+
+        if (empty($contact)) {
+            $result['status'] = 0;
+            return $result;
+        }
+
         $contact->delete();
-        return redirect('/');
-    
+        return $result;
     }
 }
